stream: Streaming API for decompression
Create a new struct `Stream` that uses the `std::io::Write` interface
to read chunks of compressed data and write them to an output sink.

Add a streaming mode so processing can work with streaming chunks of
data. This is required because process() assumed the input reader
contained a complete stream.

A CheckState, check() methods and try_process_next() were added to
handle when the decompressor requests more input bytes than are
available. Data is temporarily buffered in the DecoderState if more
input bytes are required to make progress.

This commit also adds utility functions to the rangecoder for working
with streaming data.

Adds an allow_incomplete option to disable end of stream checks when
calling `finish()` on a stream. This is because some users may want
to retrieve partially decompressed data.