decode: process for streaming mode
Add a streaming mode so processing can work with streaming chunks of
data. This is required because process() assumed the input reader
contained a complete stream.

A CheckState, check() methods and try_process_next() were added to
handle when the decompressor requests more input bytes than are
available. Data is temporarily buffered in the DecoderState if more
input bytes are required to make progress.

This commit also adds utility functions to the rangecoder for working
with streaming data.